created operatorNode factory class
